requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
azure-functions==1.11.2
azure-functions-durable==1.1.5
certifi==2022.6.15
charset-normalizer==2.1.0
frozenlist==1.3.0
furl==2.1.3
idna==3.3
multidict==6.0.2
orderedmultidict==1.0.1
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
urllib3==1.26.11
yarl==1.7.2
